style(test): clean code with ktlintFormat
